+++ pixijs/src/spacetime/bindings/delete_card_tracker_reducer.ts
@@ -11,5 +11,5 @@ import {
 } from "spacetimedb";
 
 export default {
-  cardId: __t.u64(),
+  cardId: __t.u32(),
 };
