+++ pixijs/src/spacetime/bindings/event_tracker_table.ts
@@ -11,7 +11,7 @@ import {
 } from "spacetimedb";
 
 export default __t.row({
-  tileId: __t.u64().primaryKey().name("tile_id"),
-  cardId: __t.u64().name("card_id"),
+  tileId: __t.u32().primaryKey().name("tile_id"),
+  cardId: __t.u32().name("card_id"),
   createTime: __t.i64().name("create_time"),
 });
