+++ pixijs/src/spacetime/bindings/tile_tracker_table.ts
@@ -11,9 +11,9 @@ import {
 } from "spacetimedb";
 
 export default __t.row({
-  tileId: __t.u64().primaryKey().name("tile_id"),
+  tileId: __t.u32().primaryKey().name("tile_id"),
   q: __t.i32(),
   r: __t.i32(),
   z: __t.i32(),
-  linkedTileId: __t.u64().name("linked_tile_id"),
+  linkedTileId: __t.u32().name("linked_tile_id"),
 });
