+++ pixijs/src/spacetime/bindings/update_player_card_reducer.ts
@@ -11,6 +11,6 @@ import {
 } from "spacetimedb";
 
 export default {
-  playerId: __t.u64(),
-  cardId: __t.u64(),
+  playerId: __t.u32(),
+  cardId: __t.u32(),
 };
