+++ pixijs/src/spacetime/bindings/upsert_card_tracker_reducer.ts
@@ -11,8 +11,8 @@ import {
 } from "spacetimedb";
 
 export default {
-  cardId: __t.u64(),
-  linkedTileId: __t.u64(),
+  cardId: __t.u32(),
+  linkedTileId: __t.u32(),
   positionLock: __t.bool(),
   positionHold: __t.bool(),
 };
